Added the health bar in all scenes + creating interfaces for refactoring the coce
