+++ Romanian MazeRunner 2D/Assets/Scripts/IHealthUtility.cs
@@ -1,0 +1,9 @@
+﻿namespace DefaultNamespace
+{
+    public interface IHealthUtility
+    {
+        public void Damage(float health);
+        public void Hurt();
+        public void Die();
+    }
+}
